Safari-609.1.11.1/LayoutTests/accessibility/ios-simulator/treeitem-expanded-expected.txt

Projects
Reports
report-1.docx
This tests that the expanded status of a treeitem is exposed to its unignored children.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS treeitem1.isExpanded is false
PASS treeitem2.isExpanded is true
PASS successfullyParsed is true

TEST COMPLETE
